refactor(kv): rename core.ts to create-kv.ts for consistency
Standardizes naming convention across dynamic and static modules:
- dynamic/kv/core.ts → dynamic/kv/create-kv.ts
- Matches pattern: create-tables.ts, create-kv.ts, create-workspace.ts
- Also exports createKv from dynamic/index.ts for completeness
