+++ packages/epicenter/src/dynamic/index.ts
@@ -73,10 +73,19 @@ export {
 } from '../core/schema/fields/types';
 
 // ════════════════════════════════════════════════════════════════════════════
+// Y.DOC STORAGE KEYS
+// ════════════════════════════════════════════════════════════════════════════
+
+export type { KvKey, TableKey as TableKeyType } from '../core/ydoc-keys';
+// Y.Doc array key conventions (for direct Y.Doc access / custom providers)
+export { KV_KEY, TableKey } from '../core/ydoc-keys';
+
+// ════════════════════════════════════════════════════════════════════════════
 // TABLES & KV
 // ════════════════════════════════════════════════════════════════════════════
 
 // KV store (YKeyValueLww-based)
+export { createKv, type Kv } from './kv/create-kv';
 export { createKvHelper, type KvHelper } from './kv/kv-helper';
 
 // Tables API (YKeyValueLww-based row-level storage)
